refactor(session): move TokenUsage from api.ts to session-contract.ts
